remove people vom team-messages
Closes Issue #239
When BIEB delete a person from a company, over the context menu, the person now gets also deleted from the team messages, the implementation is the same as when a person signout herself from a company.
Currently is very annoying when a person get removed from the team and still get team messages and is able to reply.