Do not encode special characters in password as not necessary
Previously, the password was passed through mysql_escape_string but only
when logging in. This makes some password not work. As we only pass the
password to the md5 method which accepts any data, this is not needed.

Anyway, even our quite strongly salted MD5 should not be considered
safe and we should get rid of it better sooner than later!