Streamlined setup initialization
Everything should be setup now with just one command:

  ./scripts/start

Also added command to undo everything it does:

  ./scripts/clean

... and various other commands, see README.md for more info.

I also split the npm modules into the one which runs locally
(package.json in scripts/ dir) and the one that runs for the chat
server (package.json back in chat/ where it was before). This is
because the one in scripts/ runs on the locally installed node/npm
and the one in chat/ uses the one in the docker container. They
may be different versions.