Add ci, rejig docker config, add generated assets
This is quite a big commit based on my ci branch.

ci

works with gitlab ci now, uses docker-in-docker and
docker-compose to reuse most of the dev/test setup

rejig docker config

I moved all config into docker directories to keep
it a bit neater, there were getting a lot of files
in the root dir. It's a bit more annoying to specify
the paths, but all the scripts/* work just the same
as before.

add generated assets

I didn't want to add these, but it is tricky to
build them ready for test, as you need an app server
up and running. At some point they could be generated
by a simple script and could be removed again.
This might cause a bit of hassle when deploying to
an existing production deployment as it will have
existing asset files not from git.