Put composers's vendor dir outside of container
Previously it was inside the container as it seemed a good idea
to borrow dockers ability to rebuild image when composer.json/lock
changed.

It has a few downsides though:
- it was fiddly to update composer deps
- you cannot see the vendor libs, which is useful for
  - reading, to see what they do
  - IDE debugging to be able to step into code (apparantly)

Additionally, I changed the CI docker containers to not copy the
app in. Previously it did this as we used docker-in-docker (dind)
service, which cannot "double mount" volumes, so we had to copy
instead. We no longer used dind, so no longer an issue.