Fix RegisterCept test
Previously it closed the register dialog, to then log in on the
same page. But the selector for the close button on the dialog is
crazy, and I can't find a good way to get hold of it.

So, now we simply just load / again, and log in on a fresh page...