show fetch sign ups
show future fetch signups, and give the possibily for orga team and ambadassors to sign-out the person from a slot
sign-out from a slot only works if person is orga team or is bot in the company area.

It is not possible to sign out people from a slot, if the company is in a “foreign” area where the person is not bot, this check prevents abuse the sign-out function.

For showing the next dates, the method from the dashboard get re-used, for that i move the database method to the profile module.