Remove old unit tests
There were some "unit" tests for recently introduced methods in place
that rely on live data as well as more complicated autoloading. The
tests itself are not really valuable and have not been executed for a
long time.
After failing to get them running for about 1hour, I consider it better
to just remove them and add better ones at the time this is possible.