Add very basic CLI tests for cron commands
Call all the methods that are called via cron, just to see if they
execute at all. PHP Errors are not caught on this level of testing and
are even expected as there is interaction with external systems that is
not configured (e.g. IMAP fetch) as well as with application state that
is just wrong in tests (e.g. addind a set of known user ids to a group)