Remove autoloader in console.control
The console application is only executed in CLI execution through
run.php. Although not verifiable in tests, I expect problems as
autoloaders should be registered through spl_autoload_register. Also,
the functionality is partly already in the autoloader in setup.php. The
additional part including view components is not necessary as that
folder does not even exist.