Add scripts to run php-cs-fix fix
Two scripts, one using docker and another one the local php, both to fix
the code style, are added.

This can for example be used in a pre-commit hook. This is also the
reasoning for the local version: the docker one takes about 3 seconds to
run, while running locally happens in less than 200ms when the cache is
mostly up to date.

The local version needs PHP7 to be installed locally.

A possible speedup for the docker version is to use an already running
container. It would then be comparable to the
local version. But how to detect which running container to use? That
might take more time as well.