Only load visible messages for message dropdown
The message dropdown menu is displayed at every page. In the past, it
loaded the last message of every conversation the user has ever had
(for the test user there are 155 ones) on every page loading,
then sliced the array and only displayed the first 10 of them.

The backend function that fetches those messages from the database
(MessageModel::listConversations) already had the never used feature
to limit the queried conversations. So I made the JavaScript request
use that feature.