Marking bell notification as read before redirecting to its origin
I eventually disliked the idea of doing JavaScript on the old page
after the user has clicked on a link and the browser already loads the
new page. In Firefox, the Javascript console shows error messages when
you do that, and if the site loaded faster, it might not even work.

So I removed the HTML-href from the bell menu entry and trigger the
redirect to the bell message's origin via JavaScript in the onBellClick
function - after the notification is marked as unread.