Mark all bell notification from the same origin page as read
Multiple bell notifications from the same page could turn out to be
very annoying, as the user had to click every single one of them to
mark them as read, which he normally wouldn't do if they all refer to
the same page.

So I changed the behavior to mark not only the message you clicked as
read, but all other ones that refer to the same origin page.