Refactoring: Move socket functions from Func to new websocketTrait
Those sendSock() and sendSockMulti() functions stand for their own,
classes that need to send things over socket should not have to
initialize such a big object as Func only to call those two functions.
We could move even more of Func's functions to traits in a bigger
refactoring, as many of them have nothing to do which each other and
don't need any of Func's attributes.