Make conversation store load only 10 conversatons by default
In some situations, the conversationsstore.loadConversations-method
was still called without arguments. That caused for example a newly
arrived message to load all ever held conversations into the view
store, which is meant to only contain the last 10 of them.

Now, the view store defaults to load 10 conversations, so the explicit
call to do so in the ceated()-method of the topbar entry could be
removed.