Don't bell-notify the author of a fairteiler post
Testing the bell messages with fairteiler posts, I became affected of
the circumstance that being notified by you own messages may turn out
to be very annoying.

So I fixed this for fairteiler posts. I didn't want to do it for all
bell notifications, at least not for now, because I am not sure if all
requests to send bell notifications are send by sessions of a user
that should not get notified. But we should definitly consider
implementing it fall all bell notifications, otherwise, we would have
to implement it to all sources of bell notifications.