Check permission before posting a Fairteiler post
Testing the bell notifications with different users and Fairteilers,
I found out that users without permission to post into a Fairteiler
could do so, but Fairteiler followers didn't get notified as the
permissions got checked before notifying. I thought that I could easily
call this check before posting also, but it endet up being much
trickier.

And the solution is much dirtier than I hoped it would be, as I had to
use the DI conatainer to call the mayFairteiler function of the gateway
from the WallPostPermissions class. Scaling this up to other similar
cases where the permission logic depends on the gateway or model and
needs to be checked in the WallPostPermissions class this would mean
making the the WallPostPermissions class depend on tons of model or
gateway classes. We'll have to solve this, but I think it's so big that
it should be a different issue.