Remove hard limit from the getConversations REST action
The getConversationAction in the MessageRestController only returned
the 20 last messages, which was only documented by a comment. Now,
you can specify a limit and an offset to control which messages are
sent to you. These paramters are optional, and if unset, they default
to the old behavior of this action.