Remove unused Bell-Id-Prefixes
In some cases, bell ids where prefixed with 'b-' or 'f-'. That
information seemed like it would be never used in the current state of
the code, but it caused exceptions to be thrown and bell messages to
be not marked as unread when clicked.

This should fix exception "Query 'UPDATE `fs_foodsaver_has_bell` SET `seen` = 1 WHERE `bell_id` IN (b-42) AND `foodsaver_id` = 42' can't be prepared."