Accept arrays in criterias of Database functions
A very common use case of the deprecated "execute" function is to
execute SQL statements that contain "IN (array)"" in their WHERE clause.
They could not be described as an array criteria, as these only allowed
to connect strings to other strings.

Now you can use arrays inside a criteria, and the Database functions
will automatically use the IN-operator.

For example, this criteria:
['id' => [3, 1, 5, 7]]

will be executed like

`id` IN (3, 1, 5, 7)