Reimplement 'artificial' bells as normal ones
As described in issue #353, some bell notifications where generated when
a client retrieved bell notifications from the server.

For this commit, I tried to find all possibilities where those generated
bell notifications could change status and then triggger the status
change by adding, removing or updating bell notifications. For this,
updating functionality was added to the BellGateway, and the deleting
functions where updated to notify the client about the removal.

The only status change I could not implement, where store fetches that
get outdated. That would mean to implement something like a cronjob that
removes outdated bell notifications, that gets triggered everytime a
store fetching would have taken place, but the Bieb did not accept or
deny it in time.

We'll find a solution for that.