Fix 403 for wallpost without permissions implemented
When permission logic was not implemented, WallPostPermission's
methods returned 'false', which caused a few bugs, as permissions
currently aren't implemented for many walls.

We're going to fix this on the long run, but for now, we we'll just
return true as long as no logic is implemented and the user's logged in.