Update store fetch confirmation notifications when they expire
This is an implementation of the possibility to update bell
notifications when they expire, which was introduced with the last
commit.

The expiration time of fetch confirmation notifications is set to the
next unconfirmed fetch. If the BIEB does neither confirm nor deny it,
the bell notification will expire then.

If the fetch, that was not confirmed in time, has been the only one for
this store, the notification will be deleted then. If there are other
unconfirmed fetches from this store, that are still in the future, only
the number of unconfirmed fetches in the notification's metadata gets
updated.