Fix array handling in DB functions
This fixes some bells not getting marked as unread again while updating
if more than one person should recieve the update.

DB functions that use array criteria for their where clauses are, if
an array gets assigned to a column name, supposed to transform it in to
SQL like WHERE column IN(array_content).

My first implementation just used one parameter for the array content –
so it would be "WHERE column IN(?)" and then a parameter like
"arrayElement0, arrayElement1, arrayElement2". But of course, parameters
don't get parsed (that's why we use them). So the implementation only
worked if the array had one single element and the parameter would not
need to be parsed, at it contained no commas.

Now, the prepared query is propably built like "WHERE column IN(?,?,?)"
and then, the array elements get transfered seperately. This works.