Put new stuff into private methods (Refactoring)
The code got a little bit confusing when I added more function to it.
So I moved one foreach to single function, and I found out that there
has already been a function for the generation of the IN-placeholders,
so I used that one (I also added a little bit of documentation to it).