Remove mayFairteiler()
mayFairteiler only checked if one was follower of a fairteiler, which
does not make sense as this has no effect on your permissions. Calling
this method in infofollower() caused bell notifications not to be
sent for Fairteiler posts authored by non-followers. Calling it in
WallPostPermissions made it even impossible for non-followers to post
in fairteilers.

Fixes #467