Make PushNotificationGatway implementation agnostic
WebPush notifications are not the only notifications we want to send –
we also want to support Google Firebase notifications for the Android
app and probably Apple's push service and maybe other ones in the
future. That's why I introduced PushNotificationHandlerInterface –
implement that and be granted with the power to send notification
to wherever you want. WebPushHandler is the reference implementation.

Also, I renamed the database table and columns.