Display conversation names in push messages and refactor Message functions
Added a function getProperConversationNameForFoodsaver() in
MessageGateway which solves the problem of naming conversations in one
central place. The function convMessage implemented some of that logic
before, and uses the new function now, which makes it much smaller and
not relying on a StoreGateway and a MessageGateway any more.

Also, it was moved from Control to MessageXhr, as that was the only
implementation of Control that used the function.

Finally added some tests for MessageGateway, including ones for the new
function.