Remove Ability to delete push subscriptions via REST API
This functionality does not make sense, as the browsers will never have
the ability to identify their own push subscription, as this would
be an easy way for browser fingerprinting. So nobody would have call
this in a meaningful way, and to delete disabled subscriptions, we
have the dead subscription removal after sending a notification.