Create classes for push notification metadata
Instead of arrays, the push notification metadata is now stored in
corresponding objects. In contrast to the array, the object contains
the raw metadata without any assembled message, that is directly
passed to the PushNotificationHandler. How the actual message text
and layout is assembled from the raw metadata is now responsibility of
the handler or even the client.