Use WebSocket-based solution to check whether users are online
Instead of updating a Redis key every time a user interacts with the
server and checking if the key is older than 60 seconds, we now ask the
WebSocket server if it has any connections for a user. This is much more
reliable, as it updates immediately when the user closes the browser tab
or window, and it doesn't treat users as offline if they haven't
interacted with the server for longer than a minute.

The old implementations in Mem userOnline() and userIsActive() (which
did do exactly the same thing btw.), which are now replaced with the
WebSocket solution, have been used for E-Mail notifications for missed
chat messages, E-Mails for basket requests and the online information
on the profile page. The new solution will also be used for push
notifications.

Also in this commit:
- Refactoring of MessageXhr::sendmsg() (plus sending proper status
  codes, although there is no proper error handling yet)
- Move Control::mailMessage() along with its dependencies to BasketXhr()
  (the only place where it's being used)

Kinda huge one, isn't it?