Fix green background color of first=last slot for BV
When not passed, `variant` defaults to `primary` which causes
buttons to render as btn-primary, i.e. with brown background.
Obviously, for the pickup slots we don't want that.

Also get rid of some actively harmful, otherwise useless padding.
This should be the same 2px as everywhere else, but was 5px before.