Dashboard: More translations, move some to frontend
This requires a slightly complex modification to reading the translation
key if a `source` is passed (as opposed to a `source_name`, which is not
used right now but would be rendered verbatim) -- because some logic to
determine the "best" string is still present in the backend, the key is
adjusted based on that logic (by appending `_bot` or `_own` currently).

Also shows the region (name) where an event is from now, instead of just
"Event" as source.