DO NOT USE THIS IN PRODUCTION !! Illustrate loading-filtering problem
The local seed data has lots of FSP update posts and (almost) no store
update posts. So to see the problem with almost-infinite loading:
- Open dashboard
- Scroll down => will fetch some updates
- Select filter to just display "foodsharepoints" (Fairteiler)
- Scroll down => will fetch some updates
- Select filter to just display "stores"
- Scroll down => will load itself to death, since it keeps fetching
  content (which is of type "foodsharepoint") but hiding this on
  the client (you only want to see "store" updates)

I made the three possible cases clear in code here, but the solution
is not yet clear to me. If we stop fetching as soon as a filtered
empty page happens, switching to other filter types is no longer
possible (since it looks like they won't re-enable the loading).

If you reduce the timeout to a low number of seconds, you have an
even easier time reproducing :)