Prevent trying to display event updates which don't exist
When event invitations were displayed on the dashboard, there
was a foodsaver relation that matches the DB query and so the
dashboard would try and display updates... even if none exist

If status == 3 i.e. declined, the query would fail and no such
updates would be displayed.

Out of paranoia to prevent similar incidents in the future,
add the same check to two other updatequeries (fsp, status)