Split fallBackString of abstract PushNotification into title and body
The fall back string was meant for the case that a handler or a client
does not know a specific PushNotification implementation for whatever
reason and still should be able to create a simple text notification
as a fall back.

I thought of it, and I'm sure that nearly every notification to think of
can be split into a title and a body, and even if not, nobody will be
mad at it if it returns an empty string as a body. And even the dumbest
possible PushHandler or client can show a notification based on a title
and a body, and even if it can display only a plain string it can use
only the title or concat title and body.