Port chat/websocket-server to proper TypeScript
Yes, TypeScript is a superset of JavaScript, so the chat server could be
compiled by TypeScript before. But this commit makes it compile with
"noImplicitAny": true and makes use of TypeScript's features.

Also, this commit switches to Restify as a server library and to Tedis
as a redis client. The last one was necessary because there are no
proper type bindings for the official redis, and the official redis
has no Promises/async/await, because its creator is opponent of both.

Tedis is a Redis client completely written in TypeScript, with full
support for async/await and promises.

This commit contains a breaking change, as the chat server's API, to
which our backend server talks using WebSocketConnection.php, has
been made REST conforming.

This stand of the code, especially the RestController, still needs to
be refactored, but it's a first step into a TypeScript based
architecture.