Use twig bundle instead of manually adding twig
Previously, we added the twig components we needed manually and wired
them in services.yml. TwigBundle provides a better integration of twig
into symfony as it uses a "well-known" configuration and loading
mechanism etc., other bundles might rely on it.
This Commit converts our configuration into the same for twig bundle.
Downside might be a slightly increased twig stack as we now have all
parts of TwigBundle available.