Allow basic translatability
This restructures the project a bit to comply to standard symfony
structure so we don't need to manually add translation, form and
validation service.
Then, we also follow the standard translation file name convention and
add a second, very incomplete language file for english.

Frontend is modified to be able to use a second language with fallback
as well.

A session variable is used to store the locale of a session which is
then used to configure the right locale on each request.

The REST api does not yet follow the locale settings.