Make some sense to the API of chat/webSocket server (use plural form, use commas)
Most of our REST endpoints use the plural form for resources. However,
we are not consistent with that. But as the majority uses plural (and
I personally think this makes sense), I use the plural form for "users"
here.

Also, multiple ids are now separated with commas (,) instead of dashes
(-). This is how the old API did it, and I think it kind of makes sense.
Dashes are often part of kebab-cased resource names, so separating ids
with commas could cause confusion.