Bugfix for pure-grid having some funny ideas about 100% widths
Between 959px and 48em (don't ask...) there could be dragons, depending
on loading order of CSS files. We always want 100% on dashboard and I'm
not aware of other pages with this structure, so it should be okay to
hardcode here.

36em / 48em breakpoints are taken from our outdated version of
/css/pure/grids-responsive-min.css
which ideally we would get rid of at some point