Move cache dir to var/
This is the default for symfony, we follow it by removing the special
config out of the Kernel. I did not move over our custom search index
yet as that would involve a bit of webserver reconfiguration I would to
together with moving the front controllers to a public/ dir later.