PublicProfile: first implementation
the QR codes on a foodsharing badge link to user profiles, which does
not usually help stores determine if the person is "allowed" to fetch.

the public profile section appears when you navigate to a profile page
while logged out. it shows colorful + large indicators that are easy
to understand (can / cannot pick up - only "red" for deverified users)

also there's a login button that sets the correct redirection ref,
and a motivating signup button that perhaps helps people join :)

if we decide that we want fewer pieces of information on this public
"profile" page, it can be removed in a modular way since all necessary
data is passed in using separate props

slightly flattened the profile view handling while at it