Fix pickup list bug (that has plagued Sentry) with former teammembers
Trying to display a pickup by someone who's not currently part of the
team would yield an error 500 and fail to load any of the pickup slots.

This was caused by using `getFoodsaverDetails` and assuming that it
would return all data that is mapped by `normalizeStoreUser`.

Now, use `getFoodsaver` to obtain e.g. phone numbers, and paper over the
one missing data entry (store responsibility, which cannot be determined)