1092 Add MIME type for attachments
**Why** is the change needed?

Email attachments which are retrieved via IMAP are stored
without a MIME type. This results in an incorrect `Content-Type`
header when downloading these attachments. For instance,
downloading a PDF attachment will have a `Content-Type` of
`text/html` which causes Firefox to open the file as an HTML
file instead of opening the PDF reader.

**How** is the change addressed?

When the attachment is stored, its MIME type is determined by
the `mime_content_type` function (analogously how it is done
for attachments which are uploaded via the webclient for the
`@foodsharing.network` addresses).

**Remarks**

* I first thought that we could just retrieve the MIME type
  from the email, but looking at https://github.com/ddeboer/imap/blob/45eb4913de20c0065648fe8ddf5c6427d46d0b87/src/Message/AbstractPart.php
  and https://github.com/ddeboer/imap/blob/master/src/Message/Attachment.php,
  I could not find any method that would provide this functionality.
  Therefore, I opted to reading the MIME type from the file
  after it was stored on the file system.
* There are no tests for `MailsControl.php` file yet. I didn't
  start writing a test as I believe it would require a rather
  big rewrite to make the class easier to test because in the
  current state we cannot easily inject mocks, but this may just
  be due to a lack of in-depth PHP testing knowledge.